Add Auth Manager Plus (#21)
This plugin merges nicwaller's authmgr and ianbarber's Seperate-Users plugins with 5 years worth of bug squashing, heavy code revisioning, and feature enhancement. Robust with hooks to filter data, this will integrate nicely with other plugins. I envision this as a great starting point for native user management. See the commit histories of my forks of the two plugins for a complete list of updates and enhancements.